backend/requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
yt-dlp>=2023.10.0
aiohttp>=3.9.0
pydantic>=2.0.0
sqlalchemy>=2.0.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
